Give a recording an owner
ChatRecordSession owns the state and the engine, one per button rather than a
thread-static singleton shared by every chat and story composer on the thread.
Seven booleans and an interface state become three named states, and the tri-state
stop that neither sent nor discarded becomes Cancel and Complete. Permissions go
through MediaDevicePermissions, so granting access no longer costs the first press.
