+++ Telegram/Telegram.csproj
@@ -213,6 +213,8 @@
     <Compile Include="Common\PageBlockHelper.cs" />
     <Compile Include="Common\Profiler.cs" />
     <Compile Include="Common\Recording\AudioWaveform.cs" />
+    <Compile Include="Common\Recording\ChatRecordEngine.cs" />
+    <Compile Include="Common\Recording\ChatRecordSession.cs" />
     <Compile Include="Common\Recording\VoiceSink.cs" />
     <Compile Include="Common\RichHtml.cs" />
     <Compile Include="Common\SliderHelper.cs" />
